WIP separation of concerns into different files. Current error, as we are modifying parameters inside children files/functions. Will need to overhaul logic and where variables will sit for fix
