+++ src/scripts/deleteHouse.ts
@@ -1,0 +1,7 @@
+const house = document.querySelector('#house')!;
+
+export default function deleteHouse() {
+  while (house.firstChild) {
+    house.removeChild(house.firstChild);
+  }
+}
