+++ src/scripts/getRandomInt.ts
@@ -1,0 +1,3 @@
+export default function getRandomInt(max: number) {
+  return Math.floor(Math.random() * max);
+}
